Use apache commons pool to handle connections

+++ components/org.wso2.carbon.stream.processor.core/src/main/java/org/wso2/carbon/stream/processor/core/util/BinaryEventConverter.java
@@ -69,7 +69,7 @@ public class BinaryEventConverter {
         String attributes = Arrays.toString(eventMetaInfo.getAttributeTypeOrder());
         //messageSize += eventMetaInfo.getEventSize() + attributes.length();
 
-        messageBuffer.putInt(queuedEvent.getSequenceID());
+        messageBuffer.putLong(queuedEvent.getSequenceID());
         messageBuffer.putInt((queuedEvent.getSourceHandlerElementId()).length());
         messageBuffer.put(((queuedEvent.getSourceHandlerElementId()).getBytes(Charset.defaultCharset())));
         messageBuffer.putInt((queuedEvent.getSiddhiAppName()).length());
